Update web_scraping.Rmd
Exercise 2 is a tutorial for scraping text from multiple webpages. The pages are numbered, is the page number changes in the url. This doesn't solve out problem with the Europa Parlament exampel which scrolls never endingly and only loads a number of results at a time. 

I tried to use the method in exercise 2 to collect links and articles about Afria on the Europa Parlament webpage,. I was able to identify and copy the xpath to the link, but could not  get the r code to work. I'll need your help on this. Perhaps I used the wrong xpath information or attribute name??

Any way the excerise 2 works, we need to add a short intro into how to read html, css and xpath using the inspect function in our browsers